Reformat project code, delete unused classes, add gitignore
