--- benchmark.sh
@@ -1,21 +1,0 @@
-java expr.Benchmark \
- '137' \
- 'pi' \
- 'x' \
- '3.14159 * x^2' \
- 'sin(10*x) + sin(9*x)' \
- 'sin(x) + sin(100*x)/100' \
- 'sin(0.1*x) * (sin(9*x) + sin(10*x))' \
- 'exp(-x^2)' \
- '2^(-x^2)' \
- '(x^3)^(-x^2)' \
- 'x*sin(1/x)' \
- 'x^2-x-6' \
- 'sqrt(3^2 + x^2)' \
- 'atan(5/x)' \
- '(x^2 + x + 1)/(x + 1)' \
- '(x^3 - (4*x^2) + 12)/(x^2 + 2)' \
- '-2*(x-3)^2+5' \
- '2*abs(x+1)-3' \
- 'sqrt(9-x^2)' \
- '6.674*10^-11*x*5.98*10^24/(6373002.24^2)'
