+++ com/expr/Variable.java
@@ -1,7 +1,7 @@
 // Variables associate values with names.
 // Copyright 1996 by Darius Bacon; see the file COPYING.
 
-package expr;
+package com.expr;
 
 import java.util.Hashtable;
 
@@ -10,20 +10,22 @@ import java.util.Hashtable;
  * settable value.
  */
 public class Variable extends Expr {
-    private static Hashtable variables = new Hashtable();
-    
+    private static Hashtable<String, Variable> variables = new Hashtable<>();
+
     /**
      * Return a unique variable named `name'.  There can be only one
      * variable with the same name returned by this method; that is,
      * make(s1) == make(s2) if and only if s1.equals(s2).
+     *
      * @param name the variable's name
      * @return the variable; create it initialized to 0 if it doesn't
-     *         yet exist */
+     * yet exist
+     */
     static public synchronized Variable make(String name) {
-	Variable result = (Variable) variables.get(name);
-	if (result == null)
-	    variables.put(name, result = new Variable(name));
-	return result;
+        Variable result = variables.get(name);
+        if (result == null)
+            variables.put(name, result = new Variable(name));
+        return result;
     }
 
     private String name;
@@ -31,23 +33,36 @@ public class Variable extends Expr {
 
     /**
      * Create a new variable, with initial value 0.
+     *
      * @param name the variable's name
      */
-    public Variable(String name) { 
-	this.name = name; val = 0; 
+    public Variable(String name) {
+        this.name = name;
+        val = 0;
     }
 
-    /** Return the name. */
-    public String toString() { return name; }
-
-    /** Get the value.
-     * @return the current value */
-    public double value() { 
-	return val; 
+    /**
+     * Return the name.
+     */
+    public String toString() {
+        return name;
     }
-    /** Set the value.
-     * @param value the new value */
-    public void setValue(double value) { 
-	val = value; 
+
+    /**
+     * Get the value.
+     *
+     * @return the current value
+     */
+    public double value() {
+        return val;
+    }
+
+    /**
+     * Set the value.
+     *
+     * @param value the new value
+     */
+    public void setValue(double value) {
+        val = value;
     }
 }
