--- example/Scale.java
@@ -1,4 +1,0 @@
-public abstract class Scale {
-    public abstract int map(double mathCoord);
-    public abstract double unmap(int screenCoord);
-}
