--- example/example.html
@@ -1,13 +1,0 @@
-<head>
-<title>Basic Graphing Applet</title>
-</head>
-<body>
-<h1>Basic Graphing Applet</h1>
-<hr>
-<applet code="BasicGraphApplet.class"
-	archive="expr.jar"
-	width=540 height=600
-  alt="Your browser won't run this applet even though it knows Java.">
-  If your browser knew Java and had it enabled, you'd see an applet here.
-</applet>
-</body>
